:horse: 学習処理をclassにまとめてパフォーマンスの向上 (main/client.js main/learning.js)
